Refactors to /core, /driver, and /aggregator
