+++ src/core/audits/first-meaningful-paint.js
@@ -18,7 +18,7 @@
 'use strict';
 
 const Audit = require('./audit');
-const TracingProcessor = require('../lib/traces/tracing-processor');
+const TracingProcessor = require('../../lib/traces/tracing-processor');
 
 const FAILURE_MESSAGE = 'Navigation and first paint timings not found.';
