+++ src/core/audits/input-readiness-metric.js
@@ -18,7 +18,7 @@
 'use strict';
 
 const Audit = require('./audit');
-const TracingProcessor = require('../lib/traces/tracing-processor');
+const TracingProcessor = require('../../lib/traces/tracing-processor');
 
 class InputReadinessMetric extends Audit {
   /**
