+++ src/core/audits/manifest-display.js
