+++ src/core/audits/manifest-theme-color.js
