+++ src/core/audits/screenshots.js
@@ -18,7 +18,7 @@
 'use strict';
 
 const Audit = require('./audit');
-const Formatter = require('../../formatters/formatter');
+const Formatter = require('../../../formatters/formatter');
 
 class Screenshots extends Audit {
   /**
