+++ src/core/audits/speed-index-metric.js
@@ -18,8 +18,8 @@
 'use strict';
 
 const Audit = require('./audit');
-const TracingProcessor = require('../lib/traces/tracing-processor');
-const Formatter = require('../../formatters/formatter');
+const TracingProcessor = require('../../lib/traces/tracing-processor');
+const Formatter = require('../../../formatters/formatter');
 
 // Parameters (in ms) for log-normal CDF scoring. To see the curve:
 // https://www.desmos.com/calculator/mdgjzchijg
