+++ src/core/audits/user-timings.js
@@ -18,8 +18,8 @@
 'use strict';
 
 const Audit = require('./audit');
-const Formatter = require('../../formatters/formatter');
-const TimelineModel = require('../lib/traces/devtools-timeline-model');
+const Formatter = require('../../../formatters/formatter');
+const TimelineModel = require('../../lib/traces/devtools-timeline-model');
 
 const FAILURE_MESSAGE = 'Trace data not found.';
