+++ src/core/index.js
@@ -1,0 +1,45 @@
+/**
+ * @license
+ * Copyright 2016 Google Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+'use strict';
+
+class Core {
+  static audit(artifacts, audits) {
+    return Promise.all(audits.map(audit => audit.audit(artifacts)));
+  }
+
+  static filterAndExpandAudits(audits, whitelist) {
+    return audits.filter(a => {
+          // If there is no whitelist, assume all.
+      if (!whitelist) {
+        return true;
+      }
+
+      return whitelist.has(a.toLowerCase());
+    })
+
+    // Remap the audits to its actual class.
+    .map(audit => {
+      try {
+        return require(`./audits/${audit}`);
+      } catch (requireError) {
+        throw new Error(`Unable to locate audit: ${audit}`);
+      }
+    });
+  }
+}
+
+module.exports = Core;
