+++ src/driver/gatherers/html-without-javascript.js
