+++ src/driver/gatherers/http-redirect.js
