+++ src/driver/gatherers/speedline.js
