+++ src/driver/gatherers/theme-color.js
