+++ src/lighthouse.js
@@ -16,8 +16,8 @@
  */
 'use strict';
 
-const Auditor = require('./auditor');
-const Scheduler = require('./scheduler');
+const Core = require('./core');
+const Driver = require('./driver');
 const Aggregator = require('./aggregator');
 const fs = require('fs');
 const path = require('path');
@@ -25,65 +25,7 @@ const path = require('path');
 // TODO: make this a more robust check for the config.
 function isValidConfig(config) {
   return (typeof config.passes !== 'undefined' &&
-      typeof config.audits !== 'undefined' &&
-      typeof config.aggregations);
-}
-
-function filterAndExpandAudits(audits, whitelist) {
-  return audits.filter(a => {
-        // If there is no whitelist, assume all.
-    if (!whitelist) {
-      return true;
-    }
-
-    return whitelist.has(a.toLowerCase());
-  })
-
-  // Remap the audits to its actual class.
-  .map(audit => {
-    try {
-      return require(`./audits/${audit}`);
-    } catch (requireError) {
-      throw new Error(`Unable to load audit: ${audit}`);
-    }
-  });
-}
-
-function getGatherersNeededByAudits(audits) {
-  return audits.reduce((list, audit) => {
-    audit.meta.requiredArtifacts.forEach(artifact => list.add(artifact));
-    return list;
-  }, new Set());
-}
-
-function expandPasses(audits, passes) {
-  const requiredGatherers = getGatherersNeededByAudits(audits);
-
-  return passes.map(pass => {
-    pass.gatherers = pass.gatherers
-        // Make sure we only have the gatherers that are needed by the audits
-        // that have been listed in the config.
-        .filter(gatherer => {
-          try {
-            const GathererClass = require(`./gatherers/${gatherer}`);
-            const gathererNecessary = requiredGatherers.has(GathererClass.name);
-            return gathererNecessary;
-          } catch (requireError) {
-            throw new Error(`Unable to load gatherer: ${gatherer}`);
-          }
-        })
-
-        // Take each one and instantiate it.
-        .map(gatherer => {
-          const GathererClass = require(`./gatherers/${gatherer}`);
-          return new GathererClass();
-        });
-
-    return pass;
-  })
-
-  // Make sure that any passes that have zero gatherers left are excluded from the run.
-  .filter(p => p.gatherers.length > 0);
+      typeof config.audits !== 'undefined');
 }
 
 module.exports = function(driver, opts) {
@@ -102,22 +44,29 @@ module.exports = function(driver, opts) {
     throw new Error('Config is invalid. Did you define passes, audits, and aggregations?');
   }
 
-  const audits = filterAndExpandAudits(config.audits, opts.flags.auditWhitelist);
-  const passes = expandPasses(audits, config.passes);
+  const audits = Core.filterAndExpandAudits(config.audits, opts.flags.auditWhitelist);
+  const passes = Driver.expandPasses(audits, config.passes);
 
   // The runs of Lighthouse should be tested in integration / smoke tests, so testing for coverage
   // here, at least from a unit test POV, is relatively low merit.
   /* istanbul ignore next */
-  return Scheduler
+  let run = Driver
       .run(passes, Object.assign({}, opts, {driver}))
-      .then(artifacts => Auditor.audit(artifacts, audits))
-      .then(results => Aggregator.aggregate(config.aggregations, results))
-      .then(aggregations => {
-        return {
-          url: opts.url,
-          aggregations
-        };
-      });
+      .then(artifacts => Core.audit(artifacts, audits));
+
+  // Only run aggregations if needed.
+  if (config.aggregations) {
+    run = run
+        .then(results => Aggregator.aggregate(config.aggregations, results))
+        .then(aggregations => {
+          return {
+            url: opts.url,
+            aggregations
+          };
+        });
+  }
+
+  return run;
 };
 
 /**
@@ -129,8 +78,3 @@ module.exports.getAuditList = function() {
       .readdirSync(path.join(__dirname, './audits'))
       .filter(f => /\.js$/.test(f));
 };
-
-// Expose these for testing.
-module.exports.filterAndExpandAudits = filterAndExpandAudits;
-module.exports.getGatherersNeededByAudits = getGatherersNeededByAudits;
-module.exports.expandPasses = expandPasses;
